Extract Trb SubmittedRequest component from Check
